Add slow mode AHK
Slow mode prevents a user from typing a charected within a given time of the previous key entry (typing a key resets the timer). This can be used to prevent the technicans from being able to type at a reasonable speed. Alt+s to enter slow mode Alt+f to enter fast mode (eseentially disable it by assigning the hotkeys to type the letter that is pressed).