Added Scam Breaker
This is a simple program, that runs in the background.
It will detect if a notepad is open, and then it will scan for keywords that match.. Customizable keywords
You can edit the keywords using Breaker_config.json. If it detects a keyword, it will lock the entire notepad.
And change the text inside, title of the notepad window. It takes minimal CPU.